requirements.txt
discord.py==2.4.0
asyncio==3.4.3
pillow==8.2.0
python-dotenv==0.13.0
aiohttp==3.8.1
async-timeout
requests==2.32.3
lxml==5.2.2
discord-webhook==1.3.0
feedparser==6.0.11
beautifulsoup4==4.12.3